Allow configuring support for d_ino on POSIX platforms
Some platforms that are near POSIX-compliant, but not strictly so, don't
have a `d_ino` member in `dirent`. The libc-wasi implementation
currently treats NuttX as the only exception, but there are other,
closed source platforms in the same situation.

To allow building for these platforms, we now allow the client to
specify `CONFIG_HAS_D_INO` at build time. If it's defined, we honour it;
otherwise, we keep infering it for the platforms we know.