Remove most references to __ORBIS__
Now that we've disabled hardware bounds checks on PS4, and we've added
some functionality to the POSIX compatibility library, we no longer need
most of the workarounds we had introduced in WAMR source code, so revert
them. The only remaining reference to `__ORBIS__` is for `MAP_32BIT`,
for which I'll need some more time to analyse.